Struct not_stakkr::ops::QueuedTweet

pub struct QueuedTweet {
    pub author: String,
    pub time: DateTime<FixedOffset>,
    pub content: String,
    pub time_posted: Option<DateTime<FixedOffset>>,
    pub id: Option<i64>,
}

The struct representing a queued tweet to post, posted or not.

Fields
author: String

The author to post on behalf of.

Has to have been previously added via the add-user subsystem.

time: DateTime<FixedOffset>

The time to post the tweet at.

content: String

The string content of the tweet.

time_posted: Option<DateTime<FixedOffset>>

The time this tweet was posted.

Becomes non-empty when posted.

id: Option<i64>

The numeric ID of the posted tweet.

The tweet is accessible via the standard https://twitter.com/{author}/{id} URL.

Becomes non-empty when posted.

Methods

impl QueuedTweet

fn read(p: &Path) -> Result<Vec<QueuedTweet>, Option<IoError>>

Read all queued tweets from the specified file.

fn write(queued_tweets: Vec<QueuedTweet>, p: &Path)

Save all queued tweets to the specified file.